build(warnings): stop redefining the NTSTATUS constants
pch.h supplies <windows.h> to all 565 translation units, and four of them also included
<ntstatus.h>, which redefines the same constants: 128 C4005 on a full rebuild. That volume
is what makes a genuine new warning easy to miss.

Three of the four use no STATUS_ identifier at all, so the include was dead weight. The
fourth needs STATUS_UNSUCCESSFUL as a failure sentinel and now defines it under a guard,
matching HashUtils.cpp. Two unguarded WIN32_LEAN_AND_MEAN/NOMINMAX redefinitions in a test
file accounted for the last four.

Not fixed in pch.h, because both attempts there broke the build: WIN32_NO_STATUS with
<ntstatus.h> gives 202 errors, since CryptoManager declares a constexpr STATUS_SUCCESS and
PhantomEmulator declares ~40 guest STATUS_ constants; WIN32_NO_STATUS alone gives 402, since
other units rely on winnt.h's STATUS_WAIT_0 family. A guard records both so they are not
retried.

Full rebuild: C4005 128 -> 0, C4018 2 -> 0, 0 errors. The remaining C4018 are inside
vendored gtest.h. Suite 5106/540 green, contract 724 -> 728.
Guards: NtStatusRedefinitionContractTests (4), subject derived from the tree. Mutation 4/4
exact.

Task 236 remainder.

+++ src/PhantomCore/SignatureStore/YaraRuleStore.cpp
@@ -204,7 +204,10 @@ StoreError YaraCompiler::AddFile(
         }
         
         // VALIDATION 6: File size limit (DoS protection - 10MB max per rule file)
-        if (fileSize > YaraTitaniumLimits::MAX_FILE_SIZE) {
+        // fileSize is std::streampos, MAX_FILE_SIZE is size_t. The negative case is
+        // rejected above, so the value is known non-negative here and the cast is
+        // correct rather than only quiet.
+        if (static_cast<uint64_t>(fileSize) > YaraTitaniumLimits::MAX_FILE_SIZE) {
             SS_LOG_ERROR(L"YaraCompiler", L"File too large: %s (%lld bytes)",
                 filePath.c_str(), static_cast<long long>(fileSize));
             return StoreError{SignatureStoreError::TooLarge, 0, "Rule file exceeds the maximum rule file size"};
